Fix version detection of package
https://github.com/martinbeentjes/npm-get-version-action has a bug where
it uses grep and cat to find the package version. Since our description
has the word "version" in it, it was picking up the entire description
as the version number.

I am moving the description below our package version in the file to
accommodate the bug until it can be fixed.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>