Update extension to work with Porter v1 (#58)
* Update extension to work with Porter v1

When Porter released v1, it changed the json structure of some of its commands. For example, standardizing on a naming convention (camelCase), or splitting up some commands in two like seeing an installation and its history. I have updated the extension to work with the current v1 json structure of Porter's command output.

This doesn't impact the schema (autocomplete/intellisense) when viewing a porter.yaml, that is controlled by the `porter schema` command in Porter itself actually. What this does do is fix the Porter Installation panel that is displayed on the left side of the VS Code window. You can now list installations, see previous runs, their logs and outputs again.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Explain how the namespace is defaulted when installing a bundle with the extension

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>